Added basic stats
git-svn-id: http://svn.vuze.com/public/client@2365 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus2/src/org/gudy/azureus2/core3/stats/transfer/GeneralStats.java
@@ -39,7 +39,7 @@ public interface GeneralStats {
   /**
    * @return the 'up time' in seconds
    */
-  public int getUpTime();
+  public long getUpTime();
   
   /**
    * @return the average download speed in bytes per second
